:bug: Fix broken kite install on windows 10 x64
It appears that right after the stream ended the file is still
considered in use by the system, so trying to execute it will fail. By
delaying a bit the resolution of the promise we let the time for windows
to catch up on the change and free the resource.

Ref kiteco/atom-plugin#148